backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
psycopg2-binary==2.9.10
pdfplumber==0.11.4
PyMuPDF==1.25.1
pytesseract==0.3.13
google-generativeai==0.8.4
bcrypt==4.2.1
PyJWT==2.10.1
python-dotenv==1.0.1
python-multipart==0.0.19
pydantic[email]==2.10.4
aiofiles==24.1.0
Pillow==11.1.0
